Use Synchronous authentication instead of Asynchronous.
Previous version used OpenVPN control file to do authentication. When
auth failed it wrote 1 to a control file, and 0 when auth succeeded.

On certain machines and installations, opening the control file might
be problematic. OpenVPN has a way to do synchronous authentication, and
since our code is synchronous it won't affect performance.